Move headers to include/minigame directory
