+++ include/minigame/base.h
