+++ include/minigame/mqoi.h
